Enum bcder::decode::Content

pub enum Content<'a, S: 'a> {
    Primitive(Primitive<'a, S>),
    Constructed(Constructed<'a, S>),
}

The content octets of a BER-encoded value.

A value is either primitive, containing actual octets of an actual value, or constructed, in which case its content contains additional BER encoded values. This enum is useful for cases where a certain type may be encoded as either a primitive value or a complex constructed value.

Note that this type represents the content octets only, i.e., it does not contain the tag of the value.

Variants

Primitive(Primitive<'a, S>)

The value is a primitive value.

Constructed(Constructed<'a, S>)

The value is a constructed value.

Implementations

impl<'a, S: Source + 'a> Content<'a, S>

pub fn mode(&self) -> Mode

Returns the encoding mode used by the value.

pub fn is_primitive(&self) -> bool

Returns whether this value is a primitive value.

pub fn is_constructed(&self) -> bool

Returns whether this value is a constructed value.

pub fn as_primitive( &mut self ) -> Result<&mut Primitive<'a, S>, DecodeError<S::Error>>

Converts a reference into into one to a primitive value or errors out.

pub fn as_constructed( &mut self ) -> Result<&mut Constructed<'a, S>, DecodeError<S::Error>>

Converts a reference into on to a constructed value or errors out.

pub fn content_err(&self, err: impl Into<ContentError>) -> DecodeError<S::Error>

Produces a content error at the current source position.

impl<'a, S: Source + 'a> Content<'a, S>

pub fn to_u8(&mut self) -> Result<u8, DecodeError<S::Error>>

Converts content into a u8.

If the content is not primitive or does not contain a single BER encoded INTEGER value between 0 and 256, returns a malformed error.

pub fn skip_u8_if(&mut self, expected: u8) -> Result<(), DecodeError<S::Error>>

Skips over the content if it contains an INTEGER of value expected.

The content needs to be primitive and contain a validly encoded integer of value expected or else a malformed error will be returned.

pub fn to_u16(&mut self) -> Result<u16, DecodeError<S::Error>>

Converts content into a u16.

If the content is not primitive or does not contain a single BER encoded INTEGER value between 0 and 2^16-1, returns a malformed error.

pub fn to_u32(&mut self) -> Result<u32, DecodeError<S::Error>>

Converts content into a u32.

If the content is not primitive or does not contain a single BER encoded INTEGER value between 0 and 2^32-1, returns a malformed error.

pub fn to_u64(&mut self) -> Result<u64, DecodeError<S::Error>>

Converts content into a u64.

If the content is not primitive or does not contain a single BER encoded INTEGER value between 0 and 2^64-1, returns a malformed error.

pub fn to_null(&mut self) -> Result<(), DecodeError<S::Error>>

Converts the content into a NULL value.

If the content isn’t primitive and contains a single BER encoded NULL value (i.e., nothing), returns a malformed error.