Struct bcder::string::RestrictedString

pub struct RestrictedString<L: CharSet> { /* private fields */ }

A generic restricted character string.

Restricted character strings essentially are a sequence of characters from a specific character set mapped into a sequence of octets. In BER, these are in fact encoded just like an OctetString with a different tag. Consequently, this type is a wrapper around OctetString that makes sure that the sequence of octets is correctly encoded for the given character set.

As usual, you can parse a restricted character string from encoded data by way of the take_from and from_content methods. Alternatively, you can create a new value from a String or str via the from_string and from_str associated functions.

Conversely, a restricted character string can be converted into a string by way of the to_string and to_str methods.

In addition, the restricted character string mirrors the standard library’s string types by dereffing to OctetString and using the octet string’s iterators. In addition, the chars method provides an iterator over the characters encoded in the string.

Implementations

impl<L: CharSet> RestrictedString<L>

pub fn new(os: OctetString) -> Result<Self, CharSetError>

Creates a new character string from an octet string.

If the octet string contains octet sequences that are not valid for the character set, an appropriate error will be returned.

pub fn from_string(s: String) -> Result<Self, CharSetError>

Creates a new character string from a String.

If the string’s internal representation is identical to the encoding of restricted character string, the string will be reused and no allocation occurs. Otherwise, a new bytes value is created.

If the string cannot be encoded in the character set, an error is returned.

pub fn chars(&self) -> RestrictedStringChars<'_, L>

Returns an iterator over the character in the character string.

pub fn into_bytes(self) -> Bytes

Converts the string into its underlying bytes.

Note that the bytes value will contain the raw octets of the string which are not necessarily a valid Rust string.

impl<L: CharSet> RestrictedString<L>

Decoding and Encoding

pub fn take_from<S: Source>( cons: &mut Constructed<'_, S> ) -> Result<Self, DecodeError<S::Error>>

Takes a single character set value from constructed value content.

If there is no next value, if the next value does not have the natural tag appropriate for this character set implementation, or if it does not contain a correctly encoded character string, a malformed error is returned.

pub fn from_content<S: Source>( content: &mut Content<'_, S> ) -> Result<Self, DecodeError<S::Error>>

Takes a character set from content.

pub fn encode(self) -> impl Values

Returns a value encoder for the character string with the natural tag.

pub fn encode_as(self, tag: Tag) -> impl Values

Returns a value encoder for the character string with the given tag.

pub fn encode_ref(&self) -> impl Values + '_

Returns a value encoder for the character string with the natural tag.

pub fn encode_ref_as(&self, tag: Tag) -> impl Values + '_

Returns a value encoder for the character string with the given tag.